html/browsers/history/the-location-interface/location-ancestororigins.html

<!doctype html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<iframe src=/common/blank.html></iframe>
<iframe src=/common/blank.html referrerpolicy=no-referrer></iframe>
<script>
test(() => {
  assert_equals(location.ancestorOrigins.length, 0)
}, "location.ancestorOrigins basic check")

async_test(t => {
  const frame = document.querySelector("iframe")
  frame.onload = t.step_func_done(() => {
    const ancestorOrigins = self[0].location.ancestorOrigins
    assert_equals(ancestorOrigins[0], self.origin)
    assert_equals(ancestorOrigins.length, 1)

    const rp = document.createElement("meta")
    rp.name = "referrer"
    rp.content = "no-referrer"
    document.head.appendChild(rp)
    assert_equals(self[0].location.ancestorOrigins, ancestorOrigins)
    assert_equals(ancestorOrigins[0], self.origin)
    assert_equals(ancestorOrigins.length, 1)

    frame.referrerPolicy = "no-referrer"
    assert_equals(self[0].location.ancestorOrigins, ancestorOrigins)
    assert_equals(ancestorOrigins[0], self.origin)
    assert_equals(ancestorOrigins.length, 1)
  })
}, "location.ancestorOrigins cannot be masked by a dynamic referrer policy")

async_test(t => {
  document.querySelectorAll("iframe")[1].onload = t.step_func_done(() => {
    const ancestorOrigins = self[1].location.ancestorOrigins
    assert_equals(ancestorOrigins[0], "null")
    assert_equals(ancestorOrigins.length, 1)
  })
}, "location.ancestorOrigins can be masked by a predetermined referrer policy")
</script>
